backend/scrapify/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
branca==0.6.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.6
contourpy==1.0.6
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.7
django-admin==2.0.2
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.5
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-oauth-toolkit==2.1.0
django-postgres-extensions==0.9.3
django-six==1.0.4
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.0
docutils==0.19
drf-extensions==0.7.1
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.6.0
Flask==2.2.2
folium==0.13.0
fonttools==4.38.0
h11==0.13.0
idna==3.4
imageio==2.22.2
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-examples==1.11.1
Kivy-Garden==0.1.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mysql-connector-python==8.0.31
neo4j==5.0.1
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.0
pipi==1.0.1
platformdirs==2.5.2
pluggy==1.0.0
polyline==1.4.0
protobuf==3.20.1
psycopg2==2.9.5
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-tsp==0.3.1
python3-openid==3.2.0
pytz==2022.6
PyWavelets==1.4.1
pywin32==305
requests==2.28.1
requests-oauthlib==1.3.1
scikit-image==0.19.3
scipy==1.9.3
screen==1.0.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tabulate==0.8.10
tifffile==2022.10.10
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
tsplib95==0.7.1
tzdata==2022.6
urllib3==1.26.12
vboxapi==1.0
virtualenv==20.20.0
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
xlwt==1.3.0
